separate wagmi config params from config
