+++ packages/rainbowkit/src/index.ts
@@ -2,6 +2,7 @@ export { ConnectButton } from './components/ConnectButton/ConnectButton';
 export { WalletButton } from './components/WalletButton/WalletButton';
 export { RainbowKitProvider } from './components/RainbowKitProvider/RainbowKitProvider';
 export { getDefaultConfig } from './config/getDefaultConfig';
+export { getDefaultConfigParameters } from './config/getDefaultConfigParameters';
 export { getDefaultWallets } from './wallets/getDefaultWallets';
 export { getWalletConnectConnector } from './wallets/getWalletConnectConnector';
 export { connectorsForWallets } from './wallets/connectorsForWallets';
